+++ src/gettext/upload-cli.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+const config = require('./config')
+const { uploadConfig } = require('./upload')
+
+uploadConfig(config).catch(console.error)